Handle Cooldowns Differently
ga_iOldArena was often being set -1. It's original placement in while
loop was after the LCV update.

ga_iOldArena now --'s every round without checking if it's > 0 this way
we can insure we can be sure if they were in a challenge last round by
checking if  the client ga_iCooldown is equal to the Cooldown set  value
-1.

This should fix the problem where players weren't being reset to their
old arena.